requirements.txt
cupy-cuda12x
torch
torchvision
